% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_spoon.R
\name{ask_spoon}
\alias{ask_spoon}
\title{Ask Spoon}
\usage{
ask_spoon(spoon)
}
\arguments{
\item{spoon}{A character scalar of the name of the quarto file.}
}
\description{
Render and preview a Spoon document. Note, this function is a wrapper
for the \code{quarto_preview()} function from the \code{quarto} package.
}
\examples{
\dontrun{ask_spoon("linearity")}
}
