% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.lm.R
\name{dessert.lm}
\alias{dessert.lm}
\title{\pkg{Dessert} recipes for class \code{lm}}
\usage{
dessert.lm(
  object,
  from,
  recipe,
  output_dir,
  output_format = "all",
  quiet = TRUE
)
}
\description{
\pkg{Dessert} recipe(s) for objects with class \code{lm} from the \code{stats} package.
}
\details{
\itemize{
\item ancova
\item regression
}
}
