% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.R
\name{dessert}
\alias{dessert}
\title{\pkg{Dessert} recipes}
\usage{
dessert(object, output_format = NULL, output_dir = NULL)
}
\arguments{
\item{object}{The input data set or model which forms the base of the dessert.}

\item{output_format}{The output format parsed to \code{quarto} document.}

\item{output_dir}{The output directory for the rendered output files.}
}
\value{
A reproducible standard publication archive.
}
\description{
Run a dessert recipe on a data set or model to create a reproducible standard
publication archive.
}
\examples{
\dontrun{
fit <- lm(dist ~ speed, data = cars)
dessert(fit)
}
}
