% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yamls.R
\name{get_yamls}
\alias{get_yamls}
\title{Get YAMLS}
\usage{
get_yamls(spoon)
}
\arguments{
\item{spoon}{A character vector with, at most, one element.}
}
\value{
A tibble.
}
\description{
Get YAMLS
}
\examples{
get_yamls()
}
