% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_spoon.R
\name{preview_spoon}
\alias{preview_spoon}
\alias{quit_preview}
\title{Preview Spoon}
\usage{
preview_spoon(
  spoon,
  render = "auto",
  port = "auto",
  host = "127.0.0.1",
  browse = TRUE,
  watch = TRUE,
  navigate = TRUE
)

quit_preview()
}
\arguments{
\item{spoon}{A character scalar of the name of the quarto file.}

\item{render}{For website preview, the most recent execution results of
computational documents are used to render the site (this is to optimize
startup time). If you want to perform a full render prior to serving pass
"all" or a vector of specific formats to render. Pass "default" to render
the default format for the site. For document preview, the document is
rendered prior to preview (pass \code{FALSE} to override this).}

\item{port}{Port to listen on (defaults to 4848)}

\item{host}{Hostname to bind to (defaults to 127.0.0.1)}

\item{browse}{Open a browser to preview the content. Defaults to using the
RStudio Viewer when running within RStudio.Pass a function (e.g.
\code{utils::browseURL} to override this behavior).}

\item{watch}{Watch for changes and automatically reload browser.}

\item{navigate}{Automatically navigate the preview browser to the most
recently rendered document.}
}
\description{
Render and preview a Spoon document. Note, this function is a wrapper
for the \code{quarto_preview()} function from the \code{quarto} package.
}
\examples{
\dontrun{preview_spoon("linearity")}
}
