% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_spoon.R
\name{suggest_spoon}
\alias{suggest_spoon}
\title{Suggest Spoon}
\usage{
suggest_spoon(object)
}
\arguments{
\item{object}{An input object.}
}
\description{
Suggest a spoon.
}
\examples{
\dontrun{suggest_spoon("linearity")}
}
