% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.lm.R
\name{dessert.lm}
\alias{dessert.lm}
\title{Run a dessert recipe on an analysis or data set}
\usage{
dessert.lm(object, recipe, output_dir, output_format = "all")
}
\description{
Run a dessert recipe on an analysis or data set
}
