% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.lm.R
\name{dessert.lm}
\alias{dessert.lm}
\title{\pkg{Dessert} recipes for class \code{lm}}
\usage{
dessert.lm(dessert, ...)
}
\arguments{
\item{dessert}{An dessert object with cls lm.}
}
\description{
\pkg{Dessert} recipe(s) for objects with class \code{lm} from the \code{stats} package.
}
\details{
\itemize{
\item ancova
\item regression
}
}
