% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.lm.R
\name{dessert.lm}
\alias{dessert.lm}
\title{Run a dessert recipe on an analysis or data set}
\usage{
dessert.lm(data, output_format = NULL, output_dir = NULL)
}
\arguments{
\item{data}{The input data set or fit object to call dessert on.}

\item{output_format}{The R markdown output format parsed to `rmarkdown`.}

\item{output_dir}{The output directory for the rendered output files.}
}
\value{
A reproducible standard publication archive
}
\description{
Run a dessert recipe on an analysis or data set
}
\examples{

# Call a lm dessert on a fitted linear model
fit <- lm(dist ~ speed, data = cars)
dessert.lm(fit)

# Remove the generated files in R-CMD-Check
# Don't run as an example in real life
unlink(c("dessert_envir.RData", "lm.Rmd", "lm.docx", "lm.html", "lm.pdf"))

}
