% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.lm.R
\name{dessert.lm}
\alias{dessert.lm}
\title{\pkg{Dessert} recipes for class \code{lm}}
\usage{
dessert.lm(object, from, recipe, output_dir, output_format = "all")
}
\description{
Run a \pkg{Dessert} recipe on a data set or model to create a reproducible
standard publication archive.
}
