% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.R
\name{dessert}
\alias{dessert}
\title{Run a dessert recipe on an analysis or data set}
\usage{
dessert(data, output_format = NULL, output_dir = NULL)
}
\arguments{
\item{data}{The input data set or fit object to call dessert on.}

\item{output_format}{The R markdown output format parsed to `rmarkdown`.}

\item{output_dir}{The output directory for the rendered output files.}
}
\value{
A reproducible standard publication archive
}
\description{
Run a dessert recipe on an analysis or data set
}
\examples{
fit <- lm(dist ~ speed, data = cars)
dessert(fit)
}
