% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dessert.R
\name{dessert}
\alias{dessert}
\title{\pkg{Dessert} recipes}
\usage{
dessert(
  object,
  from = NULL,
  recipe = NULL,
  output_dir = NULL,
  output_format = "all"
)
}
\arguments{
\item{object}{The input data set or model which forms the base of the dessert.}

\item{from}{A string specifying the package from which the input data set or model originates.}

\item{recipe}{A string specifying a \pkg{Dessert} recipe associated with the input data set or model. See the recipe book in details for available recipes. The default behavior is to match all desserts if no recipe is specified.}

\item{output_dir}{A string specifying the output directory for the output files.}

\item{output_format}{A string specifying the output format of the \code{quarto} document. Available ouput formats are \code{"html"}, \code{"pdf"}, and \code{"docx"}. By default, all output formats are provided.}
}
\value{
A reproducible standard publication archive with the following folder structure:
\preformatted{
| - dessert_2023-03-14_11_58_17/
  | - class_package/
    | - recipe.RData
    | - recipe.qmd
    | - recipe.html
    | - recipe.pdf
    | - recipe.docx
    | - recipe/
      | - figure_1.png
      | - figure_2.png
      | - ...
  | - ...
}
}
\description{
Run a \pkg{Dessert} recipe on a data set or model to create a reproducible
standard publication archive.
}
\details{
The \pkg{Dessert} recipe book:\tabular{llll}{
   \strong{class} \tab \strong{from} \tab \strong{recipe(s)} \tab \strong{details} \cr
   \code{lm} \tab \code{stats} \tab \code{"regression"}, \code{"ancova"} \tab \code{\link[=dessert.lm]{dessert.lm()}} \cr
}
}
\examples{
\dontrun{
fit <- lm(dist ~ speed, data = cars)
dessert(fit, recipe = "regression")
}
}
